+++ config.py
@@ -1,0 +1,4 @@
+# config.py
+BUY_PROBABILITY = 1     # Probability that a buyer will continue buying after a successful purchase
+SELLER_STOCK = 5        # Each seller starts with 5 items
+
